refactor: :sparkles: module naming for simplicity
