+++ sword/src/web/middleware/built_in/content_type.rs
@@ -1,5 +1,4 @@
-use crate::http::middleware::Next;
-use crate::web::HttpResponse;
+use crate::web::{HttpResponse, Next};
 
 use crate::{
     next,
